+++ src/test/java/in/koreatech/koin/acceptance/TimetableLectureApiTest.java
@@ -1,6 +1,5 @@
 package in.koreatech.koin.acceptance;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -20,8 +19,6 @@ import in.koreatech.koin.domain.timetable.model.Lecture;
 import in.koreatech.koin.domain.timetable.model.Semester;
 import in.koreatech.koin.domain.timetableV2.model.TimetableFrame;
 import in.koreatech.koin.domain.timetableV2.model.TimetableLecture;
-import in.koreatech.koin.domain.timetableV2.repository.TimetableFrameRepositoryV2;
-import in.koreatech.koin.domain.timetableV2.repository.TimetableLectureRepositoryV2;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.fixture.LectureFixture;
 import in.koreatech.koin.fixture.SemesterFixture;
@@ -31,7 +28,7 @@ import in.koreatech.koin.fixture.UserFixture;
 @SuppressWarnings("NonAsciiCharacters")
 @Transactional
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-public class TimetableV2ApiTest extends AcceptanceTest {
+public class TimetableLectureApiTest extends AcceptanceTest {
 
     @Autowired
     private TimeTableV2Fixture timetableV2Fixture;
@@ -45,224 +42,20 @@ public class TimetableV2ApiTest extends AcceptanceTest {
     @Autowired
     private LectureFixture lectureFixture;
 
-    @Autowired
-    private TimetableFrameRepositoryV2 timetableFrameRepositoryV2;
-
-    @Autowired
-    private TimetableLectureRepositoryV2 timetableLectureRepositoryV2;
+    private User user;
+    private String token;
+    private Semester semester;
 
     @BeforeAll
     void setup() {
         clear();
-    }
-
-    @Test
-    void 특정_시간표_frame을_생성한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
-        mockMvc.perform(
-                post("/v2/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .content(String.format("""
-                        {
-                            "semester": "%s"
-                        }
-                        """, semester.getSemester()
-                    ))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isOk())
-            .andExpect(content().json("""
-                {
-                    "id": 1,
-                    "timetable_name": "시간표1",
-                    "is_main": true
-                }
-                """));
-    }
-
-    @Test
-    void 특정_시간표_frame을_이름을_지어_생성한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
-        mockMvc.perform(
-                post("/v2/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .content(String.format("""
-                        {
-                            "semester": "%s",
-                            "timetable_name": "%s"
-                        }
-                        """, semester.getSemester(), "이름지어본시간표"
-                    ))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isOk())
-            .andExpect(content().json("""
-                {
-                    "id": 1,
-                    "timetable_name": "이름지어본시간표",
-                    "is_main": true
-                }
-                """));
-    }
-
-    @Test
-    void 특정_시간표_frame을_수정한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-        TimetableFrame frame = timetableV2Fixture.시간표1(user, semester);
-        Integer frameId = frame.getId();
-
-        mockMvc.perform(
-                put("/v2/timetables/frame/{id}", frameId)
-                    .header("Authorization", "Bearer " + token)
-                    .content(String.format("""
-                        {
-                            "timetable_name": "새로운 이름",
-                            "is_main": true
-                        }
-                        """
-                    ))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isOk())
-            .andExpect(content().json("""
-                {
-                    "id": 1,
-                    "timetable_name": "새로운 이름",
-                    "is_main": true
-                }
-                """));
-    }
-
-    @Test
-    void 모든_시간표_frame을_조회한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
-        timetableV2Fixture.시간표1(user, semester);
-        timetableV2Fixture.시간표2(user, semester);
-
-        mockMvc.perform(
-                get("/v2/timetables/frames")
-                    .header("Authorization", "Bearer " + token)
-                    .param("semester", semester.getSemester())
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isOk())
-            .andExpect(content().json("""
-                [
-                    {
-                        "id": 1,
-                        "timetable_name": "시간표1",
-                        "is_main": true
-                    },
-                    {
-                        "id": 2,
-                        "timetable_name": "시간표2",
-                        "is_main": false
-                    }
-                ]
-                """));
-    }
-
-    @Test
-    void 강의를_담고_있는_특정_시간표_frame을_삭제한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-        Lecture lecture = lectureFixture.HRD_개론(semester.getSemester());
-
-        TimetableFrame frame1 = timetableV2Fixture.시간표5(user, semester, lecture);
-
-        mockMvc.perform(
-                delete("/v2/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .param("id", String.valueOf(frame1.getId()))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isNoContent());
-
-        assertThat(timetableFrameRepositoryV2.findById(frame1.getId())).isNotPresent();
-        assertThat(timetableLectureRepositoryV2.findById(frame1.getTimetableLectures().get(1).getId())).isNotPresent();
-    }
-
-    @Test
-    void isMain인_frame을_삭제한다_다른_frame이_main으로_됨() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
-        TimetableFrame frame1 = timetableV2Fixture.시간표1(user, semester);
-        TimetableFrame frame2 = timetableV2Fixture.시간표2(user, semester);
-
-        mockMvc.perform(
-                delete("/v2/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .param("id", String.valueOf(frame1.getId()))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isNoContent());
-
-        assertThat(timetableFrameRepositoryV2.findById(frame1.getId())).isNotPresent();
-
-        TimetableFrame reloadedFrame2 = timetableFrameRepositoryV2.findById(frame2.getId()).orElseThrow();
-        assertThat(reloadedFrame2.isMain()).isTrue();
-    }
-
-    @Test
-    void 특정_시간표_frame을_삭제한다_본인_삭제가_아니면_403_반환() throws Exception {
-        User user1 = userFixture.준호_학생().getUser();
-        User user2 = userFixture.성빈_학생().getUser();
-        String token = userFixture.getToken(user2);
-        Semester semester = semesterFixture.semester("20192");
-
-        TimetableFrame frame1 = timetableV2Fixture.시간표1(user1, semester);
-
-        mockMvc.perform(
-                delete("/v2/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .param("id", String.valueOf(frame1.getId()))
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isForbidden());
-    }
-
-    @Test
-    void 모든_시간표_프레임을_삭제한다() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
-        TimetableFrame frame1 = timetableV2Fixture.시간표1(user, semester);
-        TimetableFrame frame2 = timetableV2Fixture.시간표1(user, semester);
-        TimetableFrame frame3 = timetableV2Fixture.시간표1(user, semester);
-
-        mockMvc.perform(
-                delete("/v2/all/timetables/frame")
-                    .header("Authorization", "Bearer " + token)
-                    .param("semester", semester.getSemester())
-                    .contentType(MediaType.APPLICATION_JSON)
-            )
-            .andExpect(status().isNoContent());
-
-        assertThat(timetableFrameRepositoryV2.findById(frame1.getId())).isNotPresent();
-        assertThat(timetableFrameRepositoryV2.findById(frame2.getId())).isNotPresent();
-        assertThat(timetableFrameRepositoryV2.findById(frame3.getId())).isNotPresent();
+        user = userFixture.준호_학생().getUser();
+        token = userFixture.getToken(user);
+        semester = semesterFixture.semester("20192");
     }
 
     @Test
     void 시간표를_생성한다_TimetableLecture() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
         timetableV2Fixture.시간표1(user, semester);
 
         mockMvc.perform(
@@ -339,11 +132,7 @@ public class TimetableV2ApiTest extends AcceptanceTest {
 
     @Test
     void 시간표를_수정한다_TimetableLecture() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-        TimetableFrame frame = timetableV2Fixture.시간표3(user, semester);
-        Integer frameId = frame.getId();
+        timetableV2Fixture.시간표3(user, semester);
 
         mockMvc.perform(
                 put("/v2/timetables/lecture")
@@ -421,10 +210,6 @@ public class TimetableV2ApiTest extends AcceptanceTest {
 
     @Test
     void 시간표를_조회한다_TimetableLecture() throws Exception {
-        User user = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user);
-        Semester semester = semesterFixture.semester("20192");
-
         Lecture 건축구조의_이해_및_실습 = lectureFixture.건축구조의_이해_및_실습(semester.getSemester());
         Lecture HRD_개론 = lectureFixture.HRD_개론(semester.getSemester());
 
@@ -482,12 +267,9 @@ public class TimetableV2ApiTest extends AcceptanceTest {
 
     @Test
     void 시간표에서_특정_강의를_삭제한다() throws Exception {
-        User user1 = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user1);
-        Semester semester = semesterFixture.semester("20192");
         Lecture lecture1 = lectureFixture.HRD_개론("20192");
         Lecture lecture2 = lectureFixture.영어청해("20192");
-        TimetableFrame frame = timetableV2Fixture.시간표4(user1, semester, lecture1, lecture2);
+        TimetableFrame frame = timetableV2Fixture.시간표4(user, semester, lecture1, lecture2);
 
         Integer lectureId = lecture1.getId();
 
@@ -502,12 +284,9 @@ public class TimetableV2ApiTest extends AcceptanceTest {
 
     @Test
     void 시간표에서_특정_강의를_삭제한다_V2() throws Exception {
-        User user1 = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user1);
-        Semester semester = semesterFixture.semester("20192");
         Lecture lecture1 = lectureFixture.HRD_개론("20192");
         Lecture lecture2 = lectureFixture.영어청해("20192");
-        TimetableFrame frame = timetableV2Fixture.시간표4(user1, semester, lecture1, lecture2);
+        TimetableFrame frame = timetableV2Fixture.시간표4(user, semester, lecture1, lecture2);
 
         Integer frameId = frame.getId();
         Integer lectureId = lecture1.getId();
@@ -522,12 +301,9 @@ public class TimetableV2ApiTest extends AcceptanceTest {
 
     @Test
     void 시간표에서_여러개의_강의를_한번에_삭제한다_V2() throws Exception {
-        User user1 = userFixture.준호_학생().getUser();
-        String token = userFixture.getToken(user1);
-        Semester semester = semesterFixture.semester("20192");
         Lecture lecture1 = lectureFixture.HRD_개론("20192");
         Lecture lecture2 = lectureFixture.영어청해("20192");
-        TimetableFrame frame = timetableV2Fixture.시간표4(user1, semester, lecture1, lecture2);
+        TimetableFrame frame = timetableV2Fixture.시간표4(user, semester, lecture1, lecture2);
 
         List<Integer> timetableLectureIds = frame.getTimetableLectures().stream()
             .map(TimetableLecture::getId)
